pyproject.toml
[tool.poetry]
name = "thiswirelesswatchdog"
version = "0.1.0"
description = ""
authors = ["franec94 <franec94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
